+++ packages/lumx-react/src/components/index.ts
@@ -82,6 +82,21 @@ enum Emphasis {
 }
 
 /**
+ * List of typographies.
+ */
+export enum Typography {
+    overline = 'overline',
+    caption = 'caption',
+    body1 = 'body1',
+    body2 = 'body2',
+    subtitle1 = 'subtitle1',
+    subtitle2 = 'subtitle2',
+    title = 'title',
+    headline = 'headline',
+    display1 = 'display1',
+}
+
+/**
  * All available aspect ratios.
  */
 enum AspectRatio {